Make incoming webhook accept GET, update README
Add more instructions to README, and get the `/chat` endpoint to return
something to avoid errors in the HTTP client. Switch the incoming
webhook to GET as this is more common (if not exactly better)
configuration and is account-wide so changing it for this application
could stop other applications from working.